// generated/orval/fetch/google-health-api/models/sleep.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Google Health API
 * The Google Health API lets you view and manage health and fitness metrics and measurement data.
 * OpenAPI spec version: v4
 */
import type { OutOfBedSegment } from './outOfBedSegment';
import type { SessionTimeInterval } from './sessionTimeInterval';
import type { SleepMetadata } from './sleepMetadata';
import type { SleepStage } from './sleepStage';
import type { SleepSummary } from './sleepSummary';
import type { SleepType } from './sleepType';

/**
 * A sleep session possibly including stages.
 */
export interface Sleep {
  /** Output only. Creation time of this sleep observation. */
  readonly createTime?: string;
  /** Required. Observed sleep interval. */
  interval?: SessionTimeInterval;
  /** Optional. Sleep metadata: `processed`, `main_sleep`, `manually_edited`, and `stages_status`. */
  metadata?: SleepMetadata;
  /** Optional. “Out of bed” segments that can overlap with sleep stages. */
  outOfBedSegments?: OutOfBedSegment[];
  /** Output only. List of short awake segments (under a set threshold) that are part of the sleep session. These can overlap with sleep stages. */
  readonly shortAwakenings?: readonly SleepStage[];
  /** Optional. List of non-overlapping contiguous sleep stage segments that cover the sleep period. */
  stages?: SleepStage[];
  /** Output only. Sleep summary: metrics and stages summary. */
  readonly summary?: SleepSummary;
  /** Optional. SleepType: classic or stages. */
  type?: SleepType;
  /** Output only. Last update time of this sleep observation. */
  readonly updateTime?: string;
}