// generated/orval/fetch/google-health-api/models/sleepStage.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Google Health API
 * The Google Health API lets you view and manage health and fitness metrics and measurement data.
 * OpenAPI spec version: v4
 */
import type { SleepStageType } from './sleepStageType';

/**
 * Sleep stage segment.
 */
export interface SleepStage {
  /** Output only. Creation time of this sleep stages segment. */
  readonly createTime?: string;
  /** Required. Sleep stage end time. */
  endTime?: string;
  /** Required. The offset of the user's local time at the end of the sleep stage relative to the Coordinated Universal Time (UTC). */
  endUtcOffset?: string;
  /** Required. Sleep stage start time. */
  startTime?: string;
  /** Required. The offset of the user's local time at the start of the sleep stage relative to the Coordinated Universal Time (UTC). */
  startUtcOffset?: string;
  /** Required. Sleep stage type: AWAKE, DEEP, REM, LIGHT etc. */
  type?: SleepStageType;
  /** Output only. Last update time of this sleep stages segment. */
  readonly updateTime?: string;
}