// generated/orval/fetch/google-health-api/models/sleepStageType.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Google Health API
 * The Google Health API lets you view and manage health and fitness metrics and measurement data.
 * OpenAPI spec version: v4
 */

/**
 * Required. Sleep stage type: AWAKE, DEEP, REM, LIGHT etc.
 */
export type SleepStageType = typeof SleepStageType[keyof typeof SleepStageType];


export const SleepStageType = {
  SLEEP_STAGE_TYPE_UNSPECIFIED: 'SLEEP_STAGE_TYPE_UNSPECIFIED',
  AWAKE: 'AWAKE',
  LIGHT: 'LIGHT',
  DEEP: 'DEEP',
  REM: 'REM',
  ASLEEP: 'ASLEEP',
  RESTLESS: 'RESTLESS',
} as const;