// generated/orval/fetch/google-health-api/models/sleepType.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Google Health API
 * The Google Health API lets you view and manage health and fitness metrics and measurement data.
 * OpenAPI spec version: v4
 */

/**
 * Optional. SleepType: classic or stages.
 */
export type SleepType = typeof SleepType[keyof typeof SleepType];


export const SleepType = {
  SLEEP_TYPE_UNSPECIFIED: 'SLEEP_TYPE_UNSPECIFIED',
  CLASSIC: 'CLASSIC',
  STAGES: 'STAGES',
} as const;